+++ packages/map/stories/map.js
@@ -22,20 +22,6 @@ class BasicMap extends Component {
         mapboxgl: Mapboxgl
       })
     );
-    mapObj.on('render', function(evt) {
-      let layers = ['country-label-lg', 'place-city-sm'];
-      layers.map(layer => {
-        mapObj.setLayoutProperty(layer, 'text-field', [
-          'format',
-          ['get', 'name_en'],
-          {
-            'font-scale': 1.2,
-            'text-font': ['literal', ['Roboto Bold']]
-          }
-        ]);
-        return null;
-      });
-    });
   };
 
   render() {
@@ -44,48 +30,6 @@ class BasicMap extends Component {
         longitude={5}
         latitude={34}
         zoom={1}
-        accessToken={token}
-        afterMapComplete={this.afterMapLoad}
-        height={91}
-        width={100}
-      />
-    );
-  }
-}
-
-class DarkMap extends Component {
-  afterMapLoad = mapObj => {
-    let navigation = new Mapboxgl.NavigationControl();
-    mapObj.addControl(navigation, 'top-left');
-    mapObj.addControl(
-      new MapboxGeocoder({
-        accessToken: Mapboxgl.accessToken,
-        mapboxgl: Mapboxgl
-      })
-    );
-    mapObj.on('render', function(evt) {
-      let layers = ['country-label-lg', 'place-city-sm'];
-      layers.map(layer => {
-        mapObj.setLayoutProperty(layer, 'text-field', [
-          'format',
-          ['get', 'name_en'],
-          {
-            'font-scale': 1.2,
-            'text-font': ['literal', ['Roboto Bold']]
-          }
-        ]);
-        return null;
-      });
-    });
-  };
-
-  render() {
-    return (
-      <Map
-        longitude={5}
-        latitude={34}
-        zoom={1}
-        mapStyle="dark-v9"
         accessToken={token}
         afterMapComplete={this.afterMapLoad}
         height={91}
@@ -112,6 +56,6 @@ export default ({ storiesOf }) => {
     .addDecorator(FullViewport)
     .add(
       'Dark',
-      withApiReadme(() => <DarkMap />)
+      withApiReadme(() => <BasicMap />)
     );
 };